Separo accion para cargar Novedades
