+++ src/acciones/cargarNovedadesAlumnosDesdeCsv.ts
@@ -1,0 +1,33 @@
+import { Client } from 'pg'
+import { readFile } from 'node:fs/promises';
+import { sqlLiteral } from '../cli'
+
+
+async function leerYParsearCsv(filePath:string){
+    const contents = await readFile(filePath, { encoding: 'utf8' });
+    const header = contents.split(/\r?\n/)[0];
+    if (header == null) throw new Error("archivo .csv vacio");
+    const columns = header.split(',').map(col => col.trim());
+    const dataLines = contents.split(/\r?\n/).slice(1).filter(line => line.trim() !== '');
+    return {dataLines, columns};
+}
+
+async function refrescarTablaAlumnos(clientDb: Client, listaDeAlumnosCompleta:string[], columnas:string[]){
+    await clientDb.query("DELETE FROM aida.alumnos");
+    for (const line of listaDeAlumnosCompleta) {
+        const values = line.split(',');
+        const query = `
+            INSERT INTO aida.alumnos (${columnas.join(', ')}) VALUES
+                (${values.map((value) => value == '' ? 'null' : sqlLiteral(value))})
+        `;
+        console.log(query)
+        const res = await clientDb.query(query)
+        console.log(res.command, res.rowCount)
+    }
+}
+
+export async function cargarNovedadesAlumnosDesdeCsv(clientDb:Client, archivoCsv:string){
+    var {dataLines: listaDeAlumnosCompleta, columns: columnas} = await leerYParsearCsv(archivoCsv)
+    await refrescarTablaAlumnos(clientDb, listaDeAlumnosCompleta, columnas);
+}
+
